refactor: move top bar inside components folder
